+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.12.0 (Unreleased)
+## 0.11.3 (Unreleased)
 
 ### 🚨 Breaking changes
 
@@ -13,10 +13,20 @@
 ### 🛠 Improvements
 
 - [xxxx](https://github.com/vegaprotocol/vegawallet-desktop/issues/xxxx) -
-
 ### 🐛 Fixes
 
 - [xxxx](https://github.com/vegaprotocol/vegawallet-desktop/issues/xxxx) -
+
+## 0.11.2
+
+### 🛠 Improvements
+
+- [576](https://github.com/vegaprotocol/vegawallet-desktop/issues/576) - Support different vega home for Mainnet and Testnet wallet
+- [581](https://github.com/vegaprotocol/vegawallet-desktop/issues/581) - Add version of the sotfware in the window title.
+
+### 🐛 Fixes
+
+- [573](https://github.com/vegaprotocol/vegawallet-desktop/issues/573) - Indirectly fix `ERR_FILE_NOT_FOUND` error by separated Fairground and Mainnet software, so configuration files are not corrupted any more.
 
 ## 0.11.1